requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
python-dotenv
python-multipart
pytest
httpx